search

+++ src/main/webapp/view/ctsp/get-all.jsp
@@ -1,5 +1,6 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/fmt" prefix="fmt" %>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8" %>
 <!doctype html>
 <html lang="en">
 <head>
@@ -13,15 +14,53 @@
     <title>Document</title>
 </head>
 
-<body>
+<body class="container mt-3">
 
 <h1>Danh sach ctsp</h1>
+
+<!-- Form tìm kiếm -->
+<%--<form action="/ctsp/search" method="get" class="form-inline mb-4">--%>
+<%--    <div class="form-group">--%>
+<%--        <label for="tenSanPham">Tên Sản phẩm:</label>--%>
+<%--        <input type="text" id="tenSanPham" name="tenSanPham" class="form-control" placeholder="Tên Sản phẩm" required>--%>
+<%--    </div>--%>
+<%--    <button type="submit" class="btn btn-primary ml-2">Tìm kiếm</button>--%>
+<%--</form>--%>
+
+<%--<form action="/ctsp/search-price" method="get">--%>
+<%--    <div class="mb-3">--%>
+<%--        <label for="minGiaBan" class="form-label">MinGiaBan</label>--%>
+<%--        <input type="number" step="1.00" class="form-control" id="minGiaBan" name="minGiaBan" required>--%>
+<%--    </div>--%>
+<%--    <div class="mb-3">--%>
+<%--        <label for="maxGiaBan" class="form-label">MaxGiaBan</label>--%>
+<%--        <input type="number" step="1.00" class="form-control" id="maxGiaBan" name="maxGiaBan" required>--%>
+<%--    </div>--%>
+<%--    <button type="submit" class="btn btn-primary">Search</button>--%>
+<%--</form>--%>
+
+<form action="/ctsp/search" method="get" class="mb-3">
+    <div class="row">
+        <div class="col-md-4">
+            <input type="text" name="tenSanPham" class="form-control" placeholder="Tên sản phẩm" value="${ctsp.tenSanPham}">
+        </div>
+        <div class="col-md-3">
+            <input type="number" name="minGiaBan" class="form-control" placeholder="Giá bán từ" value="${ctsp.minGiaBan}">
+        </div>
+        <div class="col-md-3">
+            <input type="number" name="maxGiaBan" class="form-control" placeholder="Giá bán đến" value="${ctsp.maxGiaBan}">
+        </div>
+        <div class="col-md-2">
+            <button type="submit" class="btn btn-primary">Tìm kiếm</button>
+        </div>
+    </div>
+</form>
+
 <button class="btn btn-primary"><a class="btn btn-primary" href="/ctsp/view-add">ADD</a></button>
 <table class="table table-secondary">
     <tr>
         <th>idSanPham</th>
         <th>idMauSac</th>
-        <th>idDotGiamGia</th>
         <th>anh</th>
         <th>moTa</th>
         <th>giaBan</th>
@@ -38,8 +77,8 @@
         <tr>
             <td>${ctsp.idSanPham.tenSanPham}</td>
             <td>${ctsp.idMauSac.tenMauSac}</td>
-            <td>${ctsp.idDotGiamGia.tenDotGiamGia}</td>
-            <td>${ctsp.anh.idAnh}</td>
+            <td>${ctsp.idAnh.idAnh}</td>
+<%--            <td><img src="${ctsp.idAnh.url}" alt="Anh" width="50"></td>--%>
             <td>${ctsp.moTa}</td>
             <td>${ctsp.giaBan}</td>
             <td>${ctsp.nguoiTao.email}</td>
@@ -55,12 +94,16 @@
             <td>${ctsp.ghiChu}</td>
             <td>${ctsp.trangThai}</td>
             <td>
-                <button class="btn btn-light" ><a class="btn btn-light" href="/ctsp/delete/${ctsp.idCTSP}">Delete</a></button>
+                <button class="btn btn-light" onclick="return confirm('Bạn có muốn xóa không?')" ><a class="btn btn-light" href="/ctsp/delete/${ctsp.idCTSP}">Delete</a></button>
                 <button class="btn btn-dark"><a class="btn btn-dark" href="/ctsp/updateForm/${ctsp.idCTSP}">Detail</a></button>
             </td>
         </tr>
     </c:forEach>
     </tbody>
+
+</table>
+
+<footer>
     <nav aria-label="Page navigation example">
         <ul class="pagination">
             <li class="page-item"><a class="page-link" href="/ctsp/hien-thi?page=${list.number - 1}">Truoc</a></li>
@@ -68,8 +111,5 @@
             <li class="page-item"><a class="page-link" href="/ctsp/hien-thi?page=${list.number + 1}">Sau</a></li>
         </ul>
     </nav>
-</table>
-<footer>
-
 </footer>
 </body>
